+++ source-code/clj/local_db/config.clj
@@ -24,4 +24,4 @@
 (def MAX-FILESIZE (io/MB->B 10))
 
 ; @constant (string)
-(def LOCAL-DB-PATH "monoset-environment/db/")
+(def LOCAL-DB-PATH "environment/db/")