+++ source-code/clj/x/server_core/config_handler/config.clj
@@ -20,7 +20,7 @@
 ;; ----------------------------------------------------------------------------
 
 ; @constant (string)
-(def APP-CONFIG-FILEPATH "monoset-environment/x.app-config.edn")
+(def APP-CONFIG-FILEPATH "environment/x.app-config.edn")
 
 ; @constant (string)
-(def SERVER-CONFIG-FILEPATH "monoset-environment/x.server-config.edn")
+(def SERVER-CONFIG-FILEPATH "environment/x.server-config.edn")