+++ source-code/cljs/layouts/surface_a/views.cljs
@@ -32,12 +32,9 @@
   [:div#surface-a--background [:svg {:style {:width "100%" :height "100%"}
                                      :preserve-aspect-ratio "none"
                                      :view-box              "0 0 100 100"}
-                                    [:polygon {:points "0,10 0,30 5,75 0,100 10,100 10,70 0,10 5,0 0,0"
-                                               :style  {:fill "var( --app-background-pattern-color )"}}]
-                                    [:polygon {:points "100,90 100,50 95,80 90,100 95,100"
-                                               :style  {:fill "var( --app-background-pattern-color )"}}]
-                                    [:polygon {:points "100,0 100,10 95,5 80,0"
-                                               :style  {:fill "var( --app-background-pattern-color )"}}]]])
+                                    [:polygon.surface-a--background-pattern {:points "0,10 0,30 5,75 0,100 10,100 10,70 0,10 5,0 0,0"}]
+                                    [:polygon.surface-a--background-pattern {:points "100,90 100,50 95,80 90,100 95,100"}]
+                                    [:polygon.surface-a--background-pattern {:points "100,0 100,10 95,5 80,0"}]]])